Work in progress

+++ src/gravity/Default/gravity_part.h
@@ -47,7 +47,7 @@ struct gpart {
   /* float tx; */
   /* float tv; */
 
-  float mass_interacted;
+  double mass_interacted;
   
   /* Anonymous union for id/part. */
   union {
